refactor: :recycle: Rename models files to <model>_entity.dart.
