+++ lib/src/data/infrastructure/persistance/sqlite/repository/category/sqlite_category_repository.dart
@@ -2,7 +2,7 @@ import 'package:flash_focus/src/core/identifier/identifier.dart';
 import 'package:flash_focus/src/core/persistance/repository_option.dart';
 import 'package:flash_focus/src/data/infrastructure/persistance/sqlite/entity/category/sqlite_category.dart';
 import 'package:flash_focus/src/data/infrastructure/persistance/sqlite/entity/category/sqlite_category_mapper.dart';
-import 'package:flash_focus/src/domain/category/entity/category.dart';
+import 'package:flash_focus/src/domain/category/entity/category_entity.dart';
 import 'package:flash_focus/src/domain/category/port/persistance/category_repository_port.dart';
 import 'package:sqflite/sqflite.dart';
