+++ lib/src/data/services/attempt/usecase/create_attempt_service.dart
@@ -1,4 +1,4 @@
-import 'package:flash_focus/src/domain/attempt/entity/attempt.dart';
+import 'package:flash_focus/src/domain/attempt/entity/attempt_entity.dart';
 import 'package:flash_focus/src/domain/attempt/port/persistance/attempt_repository_port.dart';
 import 'package:flash_focus/src/domain/attempt/port/usecase/create_attempt_port.dart';
 import 'package:flash_focus/src/domain/attempt/usecase/create_attempt_usecase.dart';
