+++ lib/src/domain/attempt/entity/attempt_entity.dart
