+++ lib/src/domain/attempt/port/persistance/attempt_repository_port.dart
@@ -1,6 +1,6 @@
 import 'package:flash_focus/src/core/identifier/identifier.dart';
 import 'package:flash_focus/src/core/persistance/repository_option.dart';
-import 'package:flash_focus/src/domain/attempt/entity/attempt.dart';
+import 'package:flash_focus/src/domain/attempt/entity/attempt_entity.dart';
 
 abstract class AttemptRepositoryPort {
   Future<List<AttemptEntity>> findAttempts({
