+++ lib/src/domain/attempt/port/usecase/create_attempt_port.dart
@@ -1,5 +1,5 @@
 import 'package:flash_focus/src/core/identifier/identifier.dart';
-import 'package:flash_focus/src/domain/attempt/entity/attempt.dart';
+import 'package:flash_focus/src/domain/attempt/entity/attempt_entity.dart';
 import 'package:flutter/material.dart';
 
 @immutable
