+++ lib/src/domain/card/usecase/get_card_list_usecase.dart
@@ -1,5 +1,5 @@
 import 'package:flash_focus/src/core/usecase/usecase.dart';
-import 'package:flash_focus/src/domain/card/entity/card.dart';
+import 'package:flash_focus/src/domain/card/entity/card_entity.dart';
 import 'package:flash_focus/src/domain/card/port/usecase/get_card_list_port.dart';
 
 abstract class GetCardListUseCase extends UseCase<GetCardListPort, List<CardEntity>> {}
