+++ lib/src/domain/category/entity/category_entity.dart
