+++ lib/src/domain/category/usecase/create_category_usecase.dart
@@ -1,5 +1,5 @@
 import 'package:flash_focus/src/core/usecase/usecase.dart';
-import 'package:flash_focus/src/domain/category/entity/category.dart';
+import 'package:flash_focus/src/domain/category/entity/category_entity.dart';
 import 'package:flash_focus/src/domain/category/port/usecase/create_category_port.dart';
 
 abstract class CreateCategoryUseCase extends UseCase<CreateCategoryPort, CategoryEntity> {}
